Refactor: change delete document

+++ backend/app/repository/document.js
@@ -33,14 +33,8 @@ class Document extends PostgresDB {
   async delete(userId, documentId) {
     try {
       const client = await this.pool.connect();
-      const queryExist = `                          
-        SELECT * FROM documents WHERE id = $1 and owner = $2;
-      `;
-      const valuesExist = [documentId, userId];
-      const resultExist = await client.query(queryExist, valuesExist);
 
-      if (resultExist.rows.length == 0) {
-        const queryDelUserDocuments = `                          
+      const queryDelUserDocuments = `                          
         DELETE FROM users_documents WHERE document_id = $1 AND user_id = $2
         RETURNING document_id;
       `;
@@ -50,29 +44,28 @@ class Document extends PostgresDB {
         valuesUserDocuments
       );
 
-      return resultDelUserDocuments.rows[0].id;
-      }
-
-      const queryUserDocuments = `                          
-        DELETE FROM users_documents WHERE document_id = $1 
-        RETURNING document_id ;
+      const queryUsersDocument = `                          
+        SELECT * 
+        FROM users_documents WHERE document_id = $1 ;
       `;
-      const valuesUserDocuments = [documentId];
-      const resultUserDocuments = await client.query(
-        queryUserDocuments,
-        valuesUserDocuments
+      const valuesUsersDocuments = [documentId];
+      const resultUsersDocument = await client.query(
+        queryUsersDocument,
+        valuesUsersDocuments
       );
 
-      const query = `                          
-        DELETE FROM documents WHERE id = $1 and owner = $2
-        RETURNING id ;
-      `;
-      const values = [documentId, userId];
-      const result = await client.query(query, values);
+      if (resultUsersDocument.rows.length === 0) {
+        // Delete if anyone have the link to this document
+        const query = `                          
+          DELETE FROM documents WHERE id = $1
+          RETURNING id ;
+        `;
+        const values = [documentId];
+        const result = await client.query(query, values);
+      }
 
       client.release();
-
-      return result.rows[0].id;
+      return true;
     } catch (e) {
       console.log(e);
       throw new InternalServerError("Service temporarily unavailable");
